# src/features/booking/api/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
  baseUrl: `${string}://${string}` | (string & {});
};

export type Identifier = string;

export type OffsetDateTime = string;

export type Vehicle = {
  id: Identifier;
  customerId: Identifier;
  label: string;
  registration: string;
};

export type Dealership = {
  id: Identifier;
  name: string;
  address: string;
  timezone: string;
};

export type ServiceType = {
  id: Identifier;
  name: string;
  description: string;
  durationMinutes: number;
};

export type BookingOptions = {
  vehicles: Array<Vehicle>;
  dealerships: Array<Dealership>;
  serviceTypes: Array<ServiceType>;
};

export type AvailabilitySlot = {
  startAt: OffsetDateTime;
  endAt: OffsetDateTime;
};

export type ConfirmAppointmentRequest = {
  vehicleId: Identifier;
  dealershipId: Identifier;
  serviceTypeId: Identifier;
  startAt: OffsetDateTime;
};

export type Appointment = {
  id: Identifier;
  status: "CONFIRMED" | "CANCELLED";
  vehicle: Vehicle;
  dealership: Dealership;
  serviceType: ServiceType;
  technician: AssignedResource;
  serviceBay: AssignedResource;
  startAt: OffsetDateTime;
  endAt: OffsetDateTime;
};

export type AppointmentList = {
  appointments: Array<Appointment>;
};

export type AssignedResource = {
  id: Identifier;
  name: string;
};

export type ApiError = {
  code:
    | "VALIDATION_ERROR"
    | "RESOURCE_CONFLICT"
    | "IDEMPOTENCY_CONFLICT"
    | "NOT_FOUND"
    | "INTERNAL_ERROR"
    | "NOT_READY"
    | "METHOD_NOT_ALLOWED";
  message: string;
  requestId: string;
  details?: {
    fields?: {
      [key: string]: string;
    };
  };
};

export type Health = {
  status: "ok";
};

export type VehicleId = Identifier;

export type DealershipId = Identifier;

export type ServiceTypeId = Identifier;

export type GetBookingOptionsData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/api/v1/booking-options";
};

export type GetBookingOptionsErrors = {
  /**
   * The HTTP method is not supported by this resource
   */
  405: ApiError;
  /**
   * Unexpected server failure
   */
  500: ApiError;
};

export type GetBookingOptionsError =
  GetBookingOptionsErrors[keyof GetBookingOptionsErrors];

export type GetBookingOptionsResponses = {
  /**
   * Active vehicles, service centres and service types
   */
  200: BookingOptions;
};

export type GetBookingOptionsResponse =
  GetBookingOptionsResponses[keyof GetBookingOptionsResponses];

export type GetAvailabilityData = {
  body?: never;
  path?: never;
  query: {
    vehicleId: Identifier;
    dealershipId: Identifier;
    serviceTypeId: Identifier;
    date: string;
  };
  url: "/api/v1/availability";
};

export type GetAvailabilityErrors = {
  /**
   * Request validation failed
   */
  400: ApiError;
  /**
   * The HTTP method is not supported by this resource
   */
  405: ApiError;
  /**
   * Unexpected server failure
   */
  500: ApiError;
};

export type GetAvailabilityError =
  GetAvailabilityErrors[keyof GetAvailabilityErrors];

export type GetAvailabilityResponses = {
  /**
   * Available appointment intervals
   */
  200: {
    slots: Array<AvailabilitySlot>;
  };
};

export type GetAvailabilityResponse =
  GetAvailabilityResponses[keyof GetAvailabilityResponses];

export type ListAppointmentsData = {
  body?: never;
  path?: never;
  query?: {
    status?: "CONFIRMED" | "CANCELLED";
  };
  url: "/api/v1/appointments";
};

export type ListAppointmentsErrors = {
  /**
   * Request validation failed
   */
  400: ApiError;
  /**
   * The HTTP method is not supported by this resource
   */
  405: ApiError;
  /**
   * Unexpected server failure
   */
  500: ApiError;
};

export type ListAppointmentsError =
  ListAppointmentsErrors[keyof ListAppointmentsErrors];

export type ListAppointmentsResponses = {
  /**
   * Appointments matching the optional status filter
   */
  200: AppointmentList;
};

export type ListAppointmentsResponse =
  ListAppointmentsResponses[keyof ListAppointmentsResponses];

export type ConfirmAppointmentData = {
  body: ConfirmAppointmentRequest;
  headers: {
    "Idempotency-Key": string;
  };
  path?: never;
  query?: never;
  url: "/api/v1/appointments";
};

export type ConfirmAppointmentErrors = {
  /**
   * Request validation failed
   */
  400: ApiError;
  /**
   * The HTTP method is not supported by this resource
   */
  405: ApiError;
  /**
   * The requested interval or idempotency key conflicts
   */
  409: ApiError;
  /**
   * Unexpected server failure
   */
  500: ApiError;
};

export type ConfirmAppointmentError =
  ConfirmAppointmentErrors[keyof ConfirmAppointmentErrors];

export type ConfirmAppointmentResponses = {
  /**
   * Original appointment returned for an identical idempotent replay
   */
  200: Appointment;
  /**
   * Appointment confirmed
   */
  201: Appointment;
};

export type ConfirmAppointmentResponse =
  ConfirmAppointmentResponses[keyof ConfirmAppointmentResponses];

export type GetAppointmentData = {
  body?: never;
  path: {
    appointmentId: Identifier;
  };
  query?: never;
  url: "/api/v1/appointments/{appointmentId}";
};

export type GetAppointmentErrors = {
  /**
   * Request validation failed
   */
  400: ApiError;
  /**
   * Appointment not found
   */
  404: ApiError;
  /**
   * The HTTP method is not supported by this resource
   */
  405: ApiError;
  /**
   * Unexpected server failure
   */
  500: ApiError;
};

export type GetAppointmentError =
  GetAppointmentErrors[keyof GetAppointmentErrors];

export type GetAppointmentResponses = {
  /**
   * Confirmed appointment
   */
  200: Appointment;
};

export type GetAppointmentResponse =
  GetAppointmentResponses[keyof GetAppointmentResponses];

export type GetLivenessData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/api/v1/health/live";
};

export type GetLivenessResponses = {
  /**
   * API process is alive
   */
  200: Health;
};

export type GetLivenessResponse =
  GetLivenessResponses[keyof GetLivenessResponses];

export type GetReadinessData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/api/v1/health/ready";
};

export type GetReadinessErrors = {
  /**
   * A required dependency is unavailable
   */
  503: ApiError;
};

export type GetReadinessError = GetReadinessErrors[keyof GetReadinessErrors];

export type GetReadinessResponses = {
  /**
   * API is ready
   */
  200: Health;
};

export type GetReadinessResponse =
  GetReadinessResponses[keyof GetReadinessResponses];